@@ -1,10 +1,10 @@
 import * as React from 'react';
 import * as PropTypes from 'prop-types';
-import ApolloClient, { PureQueryOptions, ApolloError } from 'apollo-client';
+import ApolloClient, { PureQueryOptions, ApolloError, FetchPolicy } from 'apollo-client';
 import { DataProxy } from 'apollo-cache';
-const invariant = require('invariant');
+import { invariant } from 'ts-invariant';
 import { DocumentNode, GraphQLError } from 'graphql';
-const shallowEqual = require('fbjs/lib/shallowEqual');
+import shallowEqual from './utils/shallowEqual';
 
 import { OperationVariables, RefetchQueriesProviderFn } from './types';
 import { parser, DocumentType } from './parser';
@@ -35,20 +35,26 @@ export declare type MutationUpdaterFn<
   }
 > = (proxy: DataProxy, mutationResult: FetchResult<T>) => void;
 
-export declare type FetchResult<C = Record<string, any>, E = Record<string, any>> = ExecutionResult<
-  C
-> & {
+export declare type FetchResult<
+  TData = Record<string, any>,
+  C = Record<string, any>,
+  E = Record<string, any>
+> = ExecutionResult<TData> & {
   extensions?: E;
   context?: C;
 };
 
-export declare type MutationOptions<TData = any, TVariables = OperationVariables> = {
+export declare type MutationOptions<
+  TData = Record<string, any>,
+  TVariables = OperationVariables
+> = {
   variables?: TVariables;
   optimisticResponse?: TData;
   refetchQueries?: Array<string | PureQueryOptions> | RefetchQueriesProviderFn;
   awaitRefetchQueries?: boolean;
   update?: MutationUpdaterFn<TData>;
   context?: Record<string, any>;
+  fetchPolicy?: FetchPolicy;
 };
 
 export declare type MutationFn<TData = any, TVariables = OperationVariables> = (
@@ -71,6 +77,7 @@ export interface MutationProps<TData = any, TVariables = OperationVariables> {
   onCompleted?: (data: TData) => void;
   onError?: (error: ApolloError) => void;
   context?: Record<string, any>;
+  fetchPolicy?: FetchPolicy;
 }
 
 export interface MutationState<TData = any> {
@@ -92,7 +99,7 @@ class Mutation<TData = any, TVariables = OperationVariables> extends React.Compo
   MutationState<TData>
 > {
   static contextTypes = {
-    client: PropTypes.object.isRequired,
+    client: PropTypes.object,
     operations: PropTypes.object,
   };
 
@@ -109,6 +116,7 @@ class Mutation<TData = any, TVariables = OperationVariables> extends React.Compo
     children: PropTypes.func.isRequired,
     onCompleted: PropTypes.func,
     onError: PropTypes.func,
+    fetchPolicy: PropTypes.string,
   };
 
   private client: ApolloClient<any>;
@@ -171,11 +179,11 @@ class Mutation<TData = any, TVariables = OperationVariables> extends React.Compo
     const mutationId = this.generateNewMutationId();
 
     return this.mutate(options)
-      .then(response => {
+      .then((response: ExecutionResult<TData>) => {
         this.onMutationCompleted(response, mutationId);
         return response;
       })
-      .catch(e => {
+      .catch((e: ApolloError) => {
         this.onMutationError(e, mutationId);
         if (!this.props.onError) throw e;
       });
@@ -189,6 +197,7 @@ class Mutation<TData = any, TVariables = OperationVariables> extends React.Compo
       update,
       context = {},
       awaitRefetchQueries = false,
+      fetchPolicy,
     } = this.props;
     const mutateOptions = { ...options };
 
@@ -220,6 +229,7 @@ class Mutation<TData = any, TVariables = OperationVariables> extends React.Compo
       awaitRefetchQueries,
       update,
       context,
+      fetchPolicy,
       variables: mutateVariables,
       ...mutateOptions,
     });
@@ -237,9 +247,6 @@ class Mutation<TData = any, TVariables = OperationVariables> extends React.Compo
   };
 
   private onMutationCompleted = (response: ExecutionResult<TData>, mutationId: number) => {
-    if (this.hasMounted === false) {
-      return;
-    }
     const { onCompleted, ignoreResults } = this.props;
 
     const { data, errors } = response;
@@ -248,7 +255,7 @@ class Mutation<TData = any, TVariables = OperationVariables> extends React.Compo
 
     const callOncomplete = () => (onCompleted ? onCompleted(data as TData) : null);
 
-    if (this.isMostRecentMutation(mutationId) && !ignoreResults) {
+    if (this.hasMounted && this.isMostRecentMutation(mutationId) && !ignoreResults) {
       this.setState({ loading: false, data, error }, callOncomplete);
     } else {
       callOncomplete();
@@ -256,13 +263,10 @@ class Mutation<TData = any, TVariables = OperationVariables> extends React.Compo
   };
 
   private onMutationError = (error: ApolloError, mutationId: number) => {
-    if (this.hasMounted === false) {
-      return;
-    }
     const { onError } = this.props;
     const callOnError = () => (onError ? onError(error) : null);
 
-    if (this.isMostRecentMutation(mutationId)) {
+    if (this.hasMounted && this.isMostRecentMutation(mutationId)) {
       this.setState({ loading: false, error }, callOnError);
     } else {
       callOnError();
