@@ -3,8 +3,6 @@ import * as PropTypes from 'prop-types';
 import ApolloClient, {
   ObservableQuery,
   ApolloError,
-  FetchPolicy,
-  ErrorPolicy,
   ApolloQueryResult,
   NetworkStatus,
   FetchMoreOptions,
@@ -12,12 +10,13 @@ import ApolloClient, {
 } from 'apollo-client';
 import { DocumentNode } from 'graphql';
 import { ZenObservable } from 'zen-observable-ts';
-import { OperationVariables, GraphqlQueryControls } from './types';
+import { OperationVariables, QueryControls, QueryOpts } from './types';
 import { parser, DocumentType, IDocumentDefinition } from './parser';
 import { getClient } from './component-utils';
+import { RenderPromises } from './getDataFromTree';
 
-const shallowEqual = require('fbjs/lib/shallowEqual');
-const invariant = require('invariant');
+import shallowEqual from './utils/shallowEqual';
+import { invariant } from 'ts-invariant';
 
 export type ObservableQueryFields<TData, TVariables> = Pick<
   ObservableQuery<TData, TVariables>,
@@ -32,21 +31,8 @@ export type ObservableQueryFields<TData, TVariables> = Pick<
     ) => Promise<ApolloQueryResult<TData2>>);
 };
 
-function compact(obj: any) {
-  return Object.keys(obj).reduce(
-    (acc, key) => {
-      if (obj[key] !== undefined) {
-        acc[key] = obj[key];
-      }
-
-      return acc;
-    },
-    {} as any,
-  );
-}
-
 function observableQueryFields<TData, TVariables>(
-  observable: ObservableQuery<TData>,
+  observable: ObservableQuery<TData, TVariables>,
 ): ObservableQueryFields<TData, TVariables> {
   const fields = {
     variables: observable.variables,
@@ -80,27 +66,19 @@ export interface QueryResult<TData = any, TVariables = OperationVariables>
   networkStatus: NetworkStatus;
 }
 
-export interface QueryProps<TData = any, TVariables = OperationVariables> {
+export interface QueryProps<TData = any, TVariables = OperationVariables> extends QueryOpts<TVariables> {
   children: (result: QueryResult<TData, TVariables>) => React.ReactNode;
-  fetchPolicy?: FetchPolicy;
-  errorPolicy?: ErrorPolicy;
-  notifyOnNetworkStatusChange?: boolean;
-  pollInterval?: number;
   query: DocumentNode;
-  variables?: TVariables;
-  ssr?: boolean;
   displayName?: string;
   skip?: boolean;
-  client?: ApolloClient<Object>;
-  context?: Record<string, any>;
-  partialRefetch?: boolean;
-  onCompleted?: (data: TData | {}) => void;
+  onCompleted?: (data: TData) => void;
   onError?: (error: ApolloError) => void;
 }
 
 export interface QueryContext {
   client?: ApolloClient<Object>;
   operations?: Map<string, { query: DocumentNode; variables: any }>;
+  renderPromises?: RenderPromises;
 }
 
 export default class Query<TData = any, TVariables = OperationVariables> extends React.Component<
@@ -109,6 +87,7 @@ export default class Query<TData = any, TVariables = OperationVariables> extends
   static contextTypes = {
     client: PropTypes.object,
     operations: PropTypes.object,
+    renderPromises: PropTypes.object,
   };
 
   static propTypes = {
@@ -132,7 +111,7 @@ export default class Query<TData = any, TVariables = OperationVariables> extends
   // request / action storage. Note that we delete querySubscription if we
   // unsubscribe but never delete queryObservable once it is created. We
   // only delete queryObservable when we unmount the component.
-  private queryObservable?: ObservableQuery<TData> | null;
+  private queryObservable?: ObservableQuery<TData, TVariables> | null;
   private querySubscription?: ZenObservable.Subscription;
   private previousData: any = {};
   private refetcherQueue?: {
@@ -143,6 +122,7 @@ export default class Query<TData = any, TVariables = OperationVariables> extends
 
   private hasMounted: boolean = false;
   private operation?: IDocumentDefinition;
+  private lastResult: ApolloQueryResult<TData> | null = null;
 
   constructor(props: QueryProps<TData, TVariables>, context: QueryContext) {
     super(props, context);
@@ -178,6 +158,12 @@ export default class Query<TData = any, TVariables = OperationVariables> extends
       ...opts,
       fetchPolicy,
     });
+
+    // Register the SSR observable, so it can be re-used once the value comes back.
+    if (this.context && this.context.renderPromises) {
+      this.context.renderPromises.registerSSRObservable(this, observable);
+    }
+
     const result = this.queryObservable!.currentResult();
 
     return result.loading ? observable.result() : false;
@@ -187,7 +173,7 @@ export default class Query<TData = any, TVariables = OperationVariables> extends
     this.hasMounted = true;
     if (this.props.skip) return;
 
-    this.startQuerySubscription();
+    this.startQuerySubscription(true);
     if (this.refetcherQueue) {
       const { args, resolve, reject } = this.refetcherQueue;
       this.queryObservable!.refetch(args)
@@ -237,32 +223,24 @@ export default class Query<TData = any, TVariables = OperationVariables> extends
       const currentResult = this.queryObservable!.currentResult();
       const { loading, error, data } = currentResult;
       if (onCompleted && !loading && !error) {
-        onCompleted(data);
+        onCompleted(data as TData);
       } else if (onError && !loading && error) {
         onError(error);
       }
     }
   }
 
-  render() {
-    const { children } = this.props;
-    const queryResult = this.getQueryResult();
-    return children(queryResult);
+  render(): React.ReactNode {
+    const { context } = this;
+    const finish = () => this.props.children(this.getQueryResult());
+    if (context && context.renderPromises) {
+      return context.renderPromises.addQueryPromise(this, finish);
+    }
+    return finish();
   }
 
   private extractOptsFromProps(props: QueryProps<TData, TVariables>) {
-    const {
-      variables,
-      pollInterval,
-      fetchPolicy,
-      errorPolicy,
-      notifyOnNetworkStatusChange,
-      query,
-      displayName = 'Query',
-      context = {},
-    } = props;
-
-    this.operation = parser(query);
+    this.operation = parser(props.query);
 
     invariant(
       this.operation.type === DocumentType.Query,
@@ -271,23 +249,30 @@ export default class Query<TData = any, TVariables = OperationVariables> extends
       }.`,
     );
 
-    return compact({
-      variables,
-      pollInterval,
-      query,
-      fetchPolicy,
-      errorPolicy,
-      notifyOnNetworkStatusChange,
-      metadata: { reactComponent: { displayName } },
-      context,
-    });
+    const displayName = props.displayName || 'Query';
+
+    return {
+      ...props,
+      displayName,
+      context: props.context || {},
+      metadata: { reactComponent: { displayName }},
+    };
   }
 
   private initializeQueryObservable(props: QueryProps<TData, TVariables>) {
     const opts = this.extractOptsFromProps(props);
     // save for backwards compat of refetcherQueries without a recycler
     this.setOperations(opts);
-    this.queryObservable = this.client.watchQuery(opts);
+
+    // See if there is an existing observable that was used to fetch the same data and
+    // if so, use it instead since it will contain the proper queryId to fetch
+    // the result set. This is used during SSR.
+    if (this.context && this.context.renderPromises) {
+      this.queryObservable = this.context.renderPromises.getSSRObservable(this);
+    }
+    if (!this.queryObservable) {
+      this.queryObservable = this.client.watchQuery(opts);
+    }
   }
 
   private setOperations(props: QueryProps<TData, TVariables>) {
@@ -315,16 +300,45 @@ export default class Query<TData = any, TVariables = OperationVariables> extends
       .catch(() => null);
   }
 
-  private startQuerySubscription = () => {
+  private startQuerySubscription = (justMounted: boolean = false) => {
+    // When the `Query` component receives new props, or when we explicitly
+    // re-subscribe to a query using `resubscribeToQuery`, we start a new
+    // subscription in this method. To avoid un-necessary re-renders when
+    // receiving new props or re-subscribing, we track the full last
+    // observable result so it can be compared against incoming new data.
+    // We only trigger a re-render if the incoming result is different than
+    // the stored `lastResult`.
+    //
+    // It's important to note that when a component is first mounted,
+    // the `startQuerySubscription` method is also triggered. During a first
+    // mount, we don't want to store or use the last result, as we always
+    // need the first render to happen, even if there was a previous last
+    // result (which can happen when the same component is mounted, unmounted,
+    // and mounted again).
+    if (!justMounted) {
+      this.lastResult = this.queryObservable!.getLastResult();
+    }
+
     if (this.querySubscription) return;
+
     // store the initial renders worth of result
     let initial: QueryResult<TData, TVariables> | undefined = this.getQueryResult();
+
     this.querySubscription = this.queryObservable!.subscribe({
-      next: ({ data }) => {
+      next: ({ loading, networkStatus, data }) => {
         // to prevent a quick second render from the subscriber
         // we compare to see if the original started finished (from cache) and is unchanged
         if (initial && initial.networkStatus === 7 && shallowEqual(initial.data, data)) {
           initial = undefined;
+          return;
+        }
+
+        if (
+          this.lastResult &&
+          this.lastResult.loading === loading &&
+          this.lastResult.networkStatus === networkStatus &&
+          shallowEqual(this.lastResult.data, data)
+        ) {
           return;
         }
 
@@ -343,6 +357,7 @@ export default class Query<TData = any, TVariables = OperationVariables> extends
 
   private removeQuerySubscription = () => {
     if (this.querySubscription) {
+      this.lastResult = this.queryObservable!.getLastResult();
       this.querySubscription.unsubscribe();
       delete this.querySubscription;
     }
@@ -352,7 +367,8 @@ export default class Query<TData = any, TVariables = OperationVariables> extends
     this.removeQuerySubscription();
 
     const lastError = this.queryObservable!.getLastError();
-    const lastResult = this.queryObservable!.getLastResult();
+    const lastResult = this.lastResult;
+
     // If lastError is set, the observable will immediately
     // send it, causing the stream to terminate on initialization.
     // We clear everything here and restore it afterward to
@@ -449,9 +465,9 @@ export default class Query<TData = any, TVariables = OperationVariables> extends
     // always hit the network with refetch, since the components data will be
     // updated and a network request is not currently active.
     if (!this.querySubscription) {
-      const oldRefetch = (data as GraphqlQueryControls).refetch;
+      const oldRefetch = (data as QueryControls<TData, TVariables>).refetch;
 
-      (data as GraphqlQueryControls).refetch = args => {
+      (data as QueryControls<TData, TVariables>).refetch = args => {
         if (this.querySubscription) {
           return oldRefetch(args);
         } else {
