@@ -8,8 +8,8 @@ import { ZenObservable } from 'zen-observable-ts';
 import { OperationVariables } from './types';
 import { getClient } from './component-utils';
 
-const shallowEqual = require('fbjs/lib/shallowEqual');
-const invariant = require('invariant');
+import shallowEqual from './utils/shallowEqual';
+import { invariant } from 'ts-invariant';
 
 export interface SubscriptionResult<TData = any> {
   loading: boolean;
@@ -29,6 +29,7 @@ export interface SubscriptionProps<TData = any, TVariables = OperationVariables>
   shouldResubscribe?: any;
   client?: ApolloClient<Object>;
   onSubscriptionData?: (options: OnSubscriptionDataOptions<TData>) => any;
+  onSubscriptionComplete?: () => void;
   children?: (result: SubscriptionResult<TData>) => React.ReactNode;
 }
 
@@ -47,7 +48,7 @@ class Subscription<TData = any, TVariables = any> extends React.Component<
   SubscriptionState<TData>
 > {
   static contextTypes = {
-    client: PropTypes.object.isRequired,
+    client: PropTypes.object,
   };
 
   static propTypes = {
@@ -55,6 +56,7 @@ class Subscription<TData = any, TVariables = any> extends React.Component<
     variables: PropTypes.object,
     children: PropTypes.func,
     onSubscriptionData: PropTypes.func,
+    onSubscriptionComplete: PropTypes.func,
     shouldResubscribe: PropTypes.oneOfType([PropTypes.func, PropTypes.bool]),
   };
 
@@ -136,6 +138,7 @@ class Subscription<TData = any, TVariables = any> extends React.Component<
     this.querySubscription = this.queryObservable!.subscribe({
       next: this.updateCurrentData,
       error: this.updateError,
+      complete: this.completeSubscription
     });
   };
 
@@ -163,6 +166,12 @@ class Subscription<TData = any, TVariables = any> extends React.Component<
       error,
       loading: false,
     });
+  };
+
+  private completeSubscription = () => {
+    const { onSubscriptionComplete } = this.props;
+    if (onSubscriptionComplete) onSubscriptionComplete();
+    this.endSubscription();
   };
 
   private endSubscription = () => {
