@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { DocumentNode } from 'graphql';
-const hoistNonReactStatics = require('hoist-non-react-statics');
+import hoistNonReactStatics from 'hoist-non-react-statics';
 
 import { parser } from './parser';
 import { MutationOpts, OperationOption, OptionProps, MutateProps } from './types';
@@ -12,11 +12,11 @@ import {
   GraphQLBase,
 } from './hoc-utils';
 
-export function mutation<
+export function withMutation<
   TProps extends TGraphQLVariables | {} = {},
   TData = {},
   TGraphQLVariables = {},
-  TChildProps = Partial<MutateProps<TData, TGraphQLVariables>>
+  TChildProps = MutateProps<TData, TGraphQLVariables>
 >(
   document: DocumentNode,
   operationOptions: OperationOption<TProps, TData, TGraphQLVariables, TChildProps> = {},
@@ -51,8 +51,6 @@ export function mutation<
           opts.variables = calculateVariablesFromProps(
             operation,
             props,
-            graphQLDisplayName,
-            getDisplayName(WrappedComponent),
           );
         }
 
