@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { ApolloError } from 'apollo-client';
 import { DocumentNode } from 'graphql';
-const hoistNonReactStatics = require('hoist-non-react-statics');
+import hoistNonReactStatics from 'hoist-non-react-statics';
 
 import { parser } from './parser';
 import { OperationOption, QueryOpts, OptionProps, DataProps } from './types';
@@ -14,11 +14,11 @@ import {
   defaultMapPropsToSkip,
 } from './hoc-utils';
 
-export function query<
+export function withQuery<
   TProps extends TGraphQLVariables | {} = {},
   TData = {},
   TGraphQLVariables = {},
-  TChildProps = Partial<DataProps<TData, TGraphQLVariables>>
+  TChildProps = DataProps<TData, TGraphQLVariables>
 >(
   document: DocumentNode,
   operationOptions: OperationOption<TProps, TData, TGraphQLVariables, TChildProps> = {},
@@ -61,8 +61,6 @@ export function query<
           opts.variables = calculateVariablesFromProps(
             operation,
             props,
-            graphQLDisplayName,
-            getDisplayName(WrappedComponent),
           );
         }
         return (
