@@ -2,9 +2,9 @@ import * as React from 'react';
 import { OperationOption } from './types';
 import ApolloConsumer from './ApolloConsumer';
 import { ApolloClient } from 'apollo-client';
+import hoistNonReactStatics from 'hoist-non-react-statics';
 
-const invariant = require('invariant');
-const hoistNonReactStatics = require('hoist-non-react-statics');
+import { invariant } from 'ts-invariant';
 
 function getDisplayName<P>(WrappedComponent: React.ComponentType<P>) {
   return WrappedComponent.displayName || WrappedComponent.name || 'Component';
