@@ -1,275 +1,158 @@
 import * as React from 'react';
+import * as PropTypes from 'prop-types';
+import Query from './Query';
+import { ObservableQuery } from 'apollo-client';
+import { DocumentNode } from 'graphql';
 
-export interface Context {
-  [key: string]: any;
+type QueryInfo = {
+  seen: boolean;
+  observable: ObservableQuery<any, any> | null;
 }
 
-interface PromiseTreeArgument {
-  rootElement: React.ReactNode;
-  rootContext: Context;
-  rootNewContext: Map<any, any>;
-}
-interface FetchComponent extends React.Component<any> {
-  fetchData(): Promise<void>;
+function makeDefaultQueryInfo(): QueryInfo {
+  return {
+    seen: false,
+    observable: null,
+  };
 }
 
-interface PromiseTreeResult {
-  promise: Promise<any>;
-  context: Context;
-  instance: FetchComponent;
-  newContext: Map<any, any>;
-}
+export class RenderPromises {
+  // Map from Query component instances to pending fetchData promises.
+  private queryPromises = new Map<Query<any, any>, Promise<any>>();
 
-interface PreactElement<P> {
-  attributes: P;
-}
+  // Two-layered map from (query document, stringified variables) to QueryInfo
+  // objects. These QueryInfo objects are intended to survive through the whole
+  // getMarkupFromTree process, whereas specific Query instances do not survive
+  // beyond a single call to renderToStaticMarkup.
+  private queryInfoTrie = new Map<DocumentNode, Map<string, QueryInfo>>();
 
-function getProps<P>(element: React.ReactElement<P> | PreactElement<P>): P {
-  return (element as React.ReactElement<P>).props || (element as PreactElement<P>).attributes;
-}
-
-function isReactElement(element: React.ReactNode): element is React.ReactElement<any> {
-  return !!(element as any).type;
-}
-
-function isComponentClass(Comp: React.ComponentType<any>): Comp is React.ComponentClass<any> {
-  return Comp.prototype && (Comp.prototype.render || Comp.prototype.isReactComponent);
-}
-
-function providesChildContext(
-  instance: React.Component<any>,
-): instance is React.Component<any> & React.ChildContextProvider<any> {
-  return !!(instance as any).getChildContext;
-}
-
-// Recurse a React Element tree, running visitor on each element.
-// If visitor returns `false`, don't call the element's render function
-// or recurse into its child elements.
-export function walkTree(
-  element: React.ReactNode,
-  context: Context,
-  visitor: (
-    element: React.ReactNode,
-    instance: React.Component<any> | null,
-    newContextMap: Map<any, any>,
-    context: Context,
-    childContext?: Context,
-  ) => boolean | void,
-  newContext: Map<any, any> = new Map(),
-) {
-  if (!element) {
-    return;
+  // Registers the server side rendered observable.
+  public registerSSRObservable<TData, TVariables>(
+    queryInstance: Query<TData, TVariables>,
+    observable: ObservableQuery<any, TVariables>,
+  ) {
+    this.lookupQueryInfo(queryInstance).observable = observable;
   }
 
-  if (Array.isArray(element)) {
-    element.forEach(item => walkTree(item, context, visitor, newContext));
-    return;
+  // Get's the cached observable that matches the SSR Query instances query and variables.
+  public getSSRObservable<TData, TVariables>(queryInstance: Query<TData, TVariables>) {
+    return this.lookupQueryInfo(queryInstance).observable;
   }
 
-  // A stateless functional component or a class
-  if (isReactElement(element)) {
-    if (typeof element.type === 'function') {
-      const Comp = element.type;
-      const props = Object.assign({}, Comp.defaultProps, getProps(element));
-      let childContext = context;
-      let child;
-
-      // Are we are a react class?
-      if (isComponentClass(Comp)) {
-        const instance = new Comp(props, context);
-        // In case the user doesn't pass these to super in the constructor.
-        // Note: `Component.props` are now readonly in `@types/react`, so
-        // we're using `defineProperty` as a workaround (for now).
-        Object.defineProperty(instance, 'props', {
-          value: instance.props || props,
-        });
-        instance.context = instance.context || context;
-
-        // Set the instance state to null (not undefined) if not set, to match React behaviour
-        instance.state = instance.state || null;
-
-        // Override setState to just change the state, not queue up an update
-        // (we can't do the default React thing as we aren't mounted
-        // "properly", however we don't need to re-render as we only support
-        // setState in componentWillMount, which happens *before* render).
-        instance.setState = newState => {
-          if (typeof newState === 'function') {
-            // React's TS type definitions don't contain context as a third parameter for
-            // setState's updater function.
-            // Remove this cast to `any` when that is fixed.
-            newState = (newState as any)(instance.state, instance.props, instance.context);
-          }
-          instance.state = Object.assign({}, instance.state, newState);
-        };
-
-        if (Comp.getDerivedStateFromProps) {
-          const result = Comp.getDerivedStateFromProps(instance.props, instance.state);
-          if (result !== null) {
-            instance.state = Object.assign({}, instance.state, result);
-          }
-        } else if (instance.UNSAFE_componentWillMount) {
-          instance.UNSAFE_componentWillMount();
-        } else if (instance.componentWillMount) {
-          instance.componentWillMount();
-        }
-
-        if (providesChildContext(instance)) {
-          childContext = Object.assign({}, context, instance.getChildContext());
-        }
-
-        if (visitor(element, instance, newContext, context, childContext) === false) {
-          return;
-        }
-
-        child = instance.render();
-      } else {
-        // Just a stateless functional
-        if (visitor(element, null, newContext, context) === false) {
-          return;
-        }
-
-        child = Comp(props, context);
-      }
-
-      if (child) {
-        if (Array.isArray(child)) {
-          child.forEach(item => walkTree(item, childContext, visitor, newContext));
-        } else {
-          walkTree(child, childContext, visitor, newContext);
-        }
-      }
-    } else if ((element.type as any)._context || (element.type as any).Consumer) {
-      // A React context provider or consumer
-      if (visitor(element, null, newContext, context) === false) {
-        return;
-      }
-
-      let child;
-      if (!!(element.type as any)._context) {
-        // A provider - sets the context value before rendering children
-        // this needs to clone the map because this value should only apply to children of the provider
-        newContext = new Map(newContext);
-        newContext.set(element.type, element.props.value);
-        child = element.props.children;
-      } else {
-        // A consumer
-        let value = (element.type as any)._currentValue;
-        if (newContext.has((element.type as any).Provider)) {
-          value = newContext.get((element.type as any).Provider);
-        }
-        child = element.props.children(value);
-      }
-
-      if (child) {
-        if (Array.isArray(child)) {
-          child.forEach(item => walkTree(item, context, visitor, newContext));
-        } else {
-          walkTree(child, context, visitor, newContext);
-        }
-      }
-    } else {
-      // A basic string or dom element, just get children
-      if (visitor(element, null, newContext, context) === false) {
-        return;
-      }
-
-      if (element.props && element.props.children) {
-        React.Children.forEach(element.props.children, (child: any) => {
-          if (child) {
-            walkTree(child, context, visitor, newContext);
-          }
-        });
-      }
-    }
-  } else if (typeof element === 'string' || typeof element === 'number') {
-    // Just visit these, they are leaves so we don't keep traversing.
-    visitor(element, null, newContext, context);
-  }
-  // TODO: Portals?
-}
-
-function hasFetchDataFunction(instance: React.Component<any>): instance is FetchComponent {
-  return typeof (instance as any).fetchData === 'function';
-}
-
-function isPromise<T>(promise: Object): promise is Promise<T> {
-  return typeof (promise as any).then === 'function';
-}
-
-function getPromisesFromTree({
-  rootElement,
-  rootContext,
-  rootNewContext,
-}: PromiseTreeArgument): PromiseTreeResult[] {
-  const promises: PromiseTreeResult[] = [];
-
-  walkTree(
-    rootElement,
-    rootContext,
-    (_, instance, newContext, context, childContext) => {
-      if (instance && hasFetchDataFunction(instance)) {
-        const promise = instance.fetchData();
-        if (isPromise<Object>(promise)) {
-          promises.push({
-            promise,
-            context: childContext || context,
-            instance,
-            newContext,
-          });
-          return false;
-        }
-      }
-    },
-    rootNewContext,
-  );
-
-  return promises;
-}
-
-function getDataAndErrorsFromTree(
-  rootElement: React.ReactNode,
-  rootContext: Object,
-  storeError: Function,
-  rootNewContext: Map<any, any> = new Map(),
-): Promise<any> {
-  const promises = getPromisesFromTree({ rootElement, rootContext, rootNewContext });
-
-  if (!promises.length) {
-    return Promise.resolve();
-  }
-
-  const mappedPromises = promises.map(({ promise, context, instance, newContext }) => {
-    return promise
-      .then(_ => getDataAndErrorsFromTree(instance.render(), context, storeError, newContext))
-      .catch(e => storeError(e));
-  });
-
-  return Promise.all(mappedPromises);
-}
-
-function processErrors(errors: any[]) {
-  switch (errors.length) {
-    case 0:
-      break;
-    case 1:
-      throw errors.pop();
-    default:
-      const wrapperError: any = new Error(
-        `${errors.length} errors were thrown when executing your fetchData functions.`,
+  public addQueryPromise<TData, TVariables>(
+    queryInstance: Query<TData, TVariables>,
+    finish: () => React.ReactNode,
+  ): React.ReactNode {
+    const info = this.lookupQueryInfo(queryInstance);
+    if (!info.seen) {
+      this.queryPromises.set(
+        queryInstance,
+        new Promise(resolve => {
+          resolve(queryInstance.fetchData());
+        }),
       );
-      wrapperError.queryErrors = errors;
-      throw wrapperError;
+      // Render null to abandon this subtree for this rendering, so that we
+      // can wait for the data to arrive.
+      return null;
+    }
+    return finish();
+  }
+
+  public hasPromises() {
+    return this.queryPromises.size > 0;
+  }
+
+  public consumeAndAwaitPromises() {
+    const promises: Promise<any>[] = [];
+    this.queryPromises.forEach((promise, queryInstance) => {
+      // Make sure we never try to call fetchData for this query document and
+      // these variables again. Since the queryInstance objects change with
+      // every rendering, deduplicating them by query and variables is the
+      // best we can do. If a different Query component happens to have the
+      // same query document and variables, it will be immediately rendered
+      // by calling finish() in addQueryPromise, which could result in the
+      // rendering of an unwanted loading state, but that's not nearly as bad
+      // as getting stuck in an infinite rendering loop because we kept calling
+      // queryInstance.fetchData for the same Query component indefinitely.
+      this.lookupQueryInfo(queryInstance).seen = true;
+      promises.push(promise);
+    });
+    this.queryPromises.clear();
+    return Promise.all(promises);
+  }
+
+  private lookupQueryInfo<TData, TVariables>(
+    queryInstance: Query<TData, TVariables>,
+  ): QueryInfo {
+    const { queryInfoTrie } = this;
+    const { query, variables } = queryInstance.props;
+    const varMap = queryInfoTrie.get(query) || new Map<string, QueryInfo>();
+    if (!queryInfoTrie.has(query)) queryInfoTrie.set(query, varMap);
+    const variablesString = JSON.stringify(variables);
+    const info = varMap.get(variablesString) || makeDefaultQueryInfo();
+    if (!varMap.has(variablesString)) varMap.set(variablesString, info);
+    return info;
   }
 }
 
 export default function getDataFromTree(
-  rootElement: React.ReactNode,
-  rootContext: any = {},
-): Promise<any> {
-  const errors: any[] = [];
-  const storeError = (error: any) => errors.push(error);
+  tree: React.ReactNode,
+  context: { [key: string]: any } = {},
+) {
+  return getMarkupFromTree({
+    tree,
+    context,
+    // If you need to configure this renderFunction, call getMarkupFromTree
+    // directly instead of getDataFromTree.
+    renderFunction: require("react-dom/server").renderToStaticMarkup,
+  });
+}
 
-  return getDataAndErrorsFromTree(rootElement, rootContext, storeError).then(_ =>
-    processErrors(errors),
-  );
+export type GetMarkupFromTreeOptions = {
+  tree: React.ReactNode;
+  context?: { [key: string]: any };
+  renderFunction?: (tree: React.ReactElement<any>) => string;
+};
+
+export function getMarkupFromTree({
+  tree,
+  context = {},
+  // The rendering function is configurable! We use renderToStaticMarkup as
+  // the default, because it's a little less expensive than renderToString,
+  // and legacy usage of getDataFromTree ignores the return value anyway.
+  renderFunction = require("react-dom/server").renderToStaticMarkup,
+}: GetMarkupFromTreeOptions): Promise<string> {
+  const renderPromises = new RenderPromises();
+
+  class RenderPromisesProvider extends React.Component {
+    static childContextTypes: { [key: string]: any } = {
+      renderPromises: PropTypes.object,
+    };
+
+    getChildContext() {
+      return { ...context, renderPromises };
+    }
+
+    render() {
+      // Always re-render from the rootElement, even though it might seem
+      // better to render the children of the component responsible for the
+      // promise, because it is not possible to reconstruct the full context
+      // of the original rendering (including all unknown context provider
+      // elements) for a subtree of the orginal component tree.
+      return tree;
+    }
+  }
+
+  Object.keys(context).forEach(key => {
+    RenderPromisesProvider.childContextTypes[key] = PropTypes.any;
+  });
+
+  function process(): Promise<string> | string {
+    const html = renderFunction(React.createElement(RenderPromisesProvider));
+    return renderPromises.hasPromises()
+      ? renderPromises.consumeAndAwaitPromises().then(process)
+      : html;
+  }
+
+  return Promise.resolve().then(process);
 }
