@@ -1,2 +1,30 @@
-export * from './browser';
+export { default as ApolloConsumer } from './ApolloConsumer';
+export * from './ApolloConsumer';
+
+export { default as ApolloProvider } from './ApolloProvider';
+export * from './ApolloProvider';
+
+export { default as Query } from './Query';
+export * from './Query';
+
+export { default as Mutation } from './Mutation';
+export * from './Mutation';
+
+export { default as Subscription } from './Subscriptions';
+export * from './Subscriptions';
+
+export { graphql } from './graphql';
+export { withQuery } from './query-hoc';
+export { withMutation } from './mutation-hoc';
+export { withSubscription } from './subscription-hoc';
+
+export { default as withApollo } from './withApollo';
+export * from './withApollo';
+
+export * from './getDataFromTree';
+export { default as getDataFromTree } from './getDataFromTree';
 export { renderToStringWithData } from './renderToStringWithData';
+
+export * from './types';
+
+export { compose } from './utils/flowRight';
