@@ -1,8 +1,9 @@
 import { ReactElement } from 'react';
-import * as ReactDOM from 'react-dom/server';
-
-import { default as getDataFromTree } from './getDataFromTree';
+import { getMarkupFromTree } from './getDataFromTree';
 
 export function renderToStringWithData(component: ReactElement<any>): Promise<string> {
-  return getDataFromTree(component).then(() => ReactDOM.renderToString(component));
+  return getMarkupFromTree({
+    tree: component,
+    renderFunction: require("react-dom/server").renderToString,
+  });
 }
