@@ -1,6 +1,6 @@
 import * as React from 'react';
 import ApolloClient from 'apollo-client';
-import { DefaultOptions } from 'apollo-client/ApolloClient';
+import { DefaultOptions, Resolvers } from 'apollo-client';
 import { InMemoryCache as Cache } from 'apollo-cache-inmemory';
 
 import { ApolloProvider } from './index';
@@ -13,6 +13,8 @@ export interface MockedProviderProps<TSerializedCache = {}> {
   addTypename?: boolean;
   defaultOptions?: DefaultOptions;
   cache?: ApolloCache<TSerializedCache>;
+  resolvers?: Resolvers;
+  childProps?: object;
 }
 
 export interface MockedProviderState {
@@ -27,30 +29,36 @@ export class MockedProvider extends React.Component<MockedProviderProps, MockedP
   constructor(props: MockedProviderProps) {
     super(props);
 
-    const { mocks, addTypename, defaultOptions, cache } = this.props;
+    const {
+      mocks,
+      addTypename,
+      defaultOptions,
+      cache,
+      resolvers,
+    } = this.props;
     const client = new ApolloClient({
       cache: cache || new Cache({ addTypename }),
       defaultOptions,
       link: new MockLink(mocks || [], addTypename),
+      resolvers,
     });
 
     this.state = { client };
   }
 
   public render() {
-    return <ApolloProvider client={this.state.client}>{this.props.children}</ApolloProvider>;
+    const { childProps } = this.props;
+
+    return (
+      <ApolloProvider client={this.state.client}>
+        {React.cloneElement(React.Children.only(this.props.children), { ...childProps })}
+      </ApolloProvider>
+    );
   }
 
   public componentWillUnmount() {
-    if (!this.state.client.queryManager) {
-      return;
-    }
-    const scheduler = this.state.client.queryManager.scheduler;
-    Object.keys(scheduler.registeredQueries).forEach(queryId => {
-      scheduler.stopPollingQuery(queryId);
-    });
-    Object.keys(scheduler.intervalQueries).forEach((interval: any) => {
-      scheduler.fetchQueriesOnInterval(interval);
-    });
+    // Since this.state.client was created in the constructor, it's this
+    // MockedProvider's responsibility to terminate it.
+    this.state.client.stop();
   }
 }
