@@ -12,6 +12,10 @@ import ApolloClient, {
 } from 'apollo-client';
 import { MutationFn } from './Mutation';
 
+export interface Context {
+  [key: string]: any;
+};
+
 export type OperationVariables = {
   [key: string]: any;
 };
@@ -31,9 +35,10 @@ export interface MutationOpts<TData = any, TGraphQLVariables = OperationVariable
   update?: MutationUpdaterFn<TData>;
   client?: ApolloClient<any>;
   notifyOnNetworkStatusChange?: boolean;
-  context?: Record<string, any>;
+  context?: Context;
   onCompleted?: (data: TData) => void;
   onError?: (error: ApolloError) => void;
+  fetchPolicy?: FetchPolicy;
 }
 
 export interface QueryOpts<TGraphQLVariables = OperationVariables> {
@@ -44,18 +49,20 @@ export interface QueryOpts<TGraphQLVariables = OperationVariables> {
   pollInterval?: number;
   client?: ApolloClient<any>;
   notifyOnNetworkStatusChange?: boolean;
-  context?: Record<string, any>;
+  context?: Context;
+  partialRefetch?: boolean;
 }
 
-export interface GraphqlQueryControls<TGraphQLVariables = OperationVariables> {
+export interface QueryControls<TData = any, TGraphQLVariables = OperationVariables> {
   error?: ApolloError;
   networkStatus: number;
   loading: boolean;
   variables: TGraphQLVariables;
   fetchMore: (
-    fetchMoreOptions: FetchMoreQueryOptions<any, any> & FetchMoreOptions,
-  ) => Promise<ApolloQueryResult<any>>;
-  refetch: (variables?: TGraphQLVariables) => Promise<ApolloQueryResult<any>>;
+    fetchMoreOptions: FetchMoreQueryOptions<TGraphQLVariables, any> &
+      FetchMoreOptions<TData, TGraphQLVariables>,
+  ) => Promise<ApolloQueryResult<TData>>;
+  refetch: (variables?: TGraphQLVariables) => Promise<ApolloQueryResult<TData>>;
   startPolling: (pollInterval: number) => void;
   stopPolling: () => void;
   subscribeToMore: (options: SubscribeToMoreOptions) => () => void;
@@ -63,12 +70,17 @@ export interface GraphqlQueryControls<TGraphQLVariables = OperationVariables> {
 }
 
 // XXX remove in the next breaking semver change (3.0)
+export interface GraphqlQueryControls<TGraphQLVariables = OperationVariables>
+  extends QueryControls<any, TGraphQLVariables> {}
+
+// XXX remove in the next breaking semver change (3.0)
 export type MutationFunc<TData = any, TVariables = OperationVariables> = MutationFn<
   TData,
   TVariables
 >;
 
-export type DataValue<TData, TGraphQLVariables = OperationVariables> = GraphqlQueryControls<
+export type DataValue<TData, TGraphQLVariables = OperationVariables> = QueryControls<
+  TData,
   TGraphQLVariables
 > &
   // data may not yet be loaded
