@@ -6,10 +6,9 @@ import { InMemoryCache as Cache } from 'apollo-cache-inmemory';
 import { ApolloLink } from 'apollo-link';
 import * as TestUtils from 'react-dom/test-utils';
 import { DocumentNode } from 'graphql';
-const compose = require('lodash.flowright');
 
 import { mockSingleLink } from '../../../src/test-utils';
-import { ApolloProvider, ChildProps, DataValue, graphql, withApollo } from '../../../src';
+import { compose, ApolloProvider, ChildProps, DataValue, graphql, withApollo } from '../../../src';
 
 describe('shared operations', () => {
   describe('withApollo', () => {
